Social Update
Overhauled the social section to include photos. Also added Twitter. A few changes were made to the links section as a whole. Some code removed in styling due to the change to images.

Body was moved to after the footer && footer now has credit for the images used for social links. Styled <a> tag.

Also changed the first sentence to about me, linked to NAD && styled <a> tag