Mobile Optimization
Previous style update to project tile width introduced a slight bug with spacing of the tiles themselves. Hopefully fixing this. Additionally I've upped the size of the nav && reversed the media query to apply to bigger screens rather than smaller so it'll shrink a little. Also messed with the border radius of all boxes (whitespace && project tile) while adding a 1px border sans color to the whitespace. May change to green later.